# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for managing home music library.",
        "title": "Home Music Library API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/song": {
            "post": {
                "description": "Создает новую песню на основе предоставленных входных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создание песни",
                "operationId": "SongCreate",
                "parameters": [
                    {
                        "description": "Входные данные для создания песни",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SongCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/song/": {
            "put": {
                "description": "Обновляет песню с таким же названием и группой",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Обновляет песню",
                "parameters": [
                    {
                        "description": "Входные данные о песне",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SongUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "description": "Удаляет песню по указанному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Удаление песни",
                "operationId": "SongDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы (обязательно).",
                        "name": "group",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название песни (обязательно).",
                        "name": "song",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/song/info": {
            "get": {
                "description": "Возвращает информацию о песни на основе входных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получение информации о песни",
                "operationId": "SongInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы (обязательно).",
                        "name": "group",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название песни (обязательно).",
                        "name": "song",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/song/search": {
            "get": {
                "description": "Поиск песни по заданным параметрам.",
                "produces": [
                    "application/json"
                ],
                "summary": "Поиск песни.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Текст встречающийся в названии группы.",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Текст встречающийся в ссылке.",
                        "name": "link",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы.",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата периода поиска песни.",
                        "name": "releaseEndDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начальная дата периода поиска песни.",
                        "name": "releaseStartDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Текст встречающийся в названии песни.",
                        "name": "song",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Текст встречающийся в тексте песни.",
                        "name": "text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/song/verses": {
            "get": {
                "description": "Получение куплетов песни с пагинацией",
                "produces": [
                    "application/json"
                ],
                "summary": "Получение куплетов песни",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы (обязательно).",
                        "name": "group",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы.",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Название песни (обязательно).",
                        "name": "song",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
        "SongCreate": {
            "type": "object",
            "required": [
                "group",
                "song"
            ],
            "properties": {
                "group": {
                    "description": "Название группы (обязательно).",
                    "type": "string"
                },
                "link": {
                    "description": "Ссылка на песню.",
                    "type": "string"
                },
                "releaseDate": {
                    "description": "Дата создания песни.",
                    "type": "string"
                },
                "song": {
                    "description": "Название песни (обязательно).",
                    "type": "string"
                },
                "text": {
                    "description": "Текст песни.",
                    "type": "string"
                }
            }
        },
        "SongUpdate": {
            "type": "object",
            "required": [
                "group",
                "song"
            ],
            "properties": {
                "group": {
                    "description": "Название группы (обязательно).",
                    "type": "string"
                },
                "link": {
                    "description": "Ссылка на песню.",
                    "type": "string"
                },
                "releaseDate": {
                    "description": "Дата создания песни.",
                    "type": "string"
                },
                "song": {
                    "description": "Название песни (обязательно).",
                    "type": "string"
                },
                "text": {
                    "description": "Текст песни.",
                    "type": "string"
                }
            }
        }
    }
}